refactor(rust): extract storage.rs with persistence helpers
